# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "LoveAC API",
        "description": "LoveACAPI API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/aac/fetch_score_info": {
            "post": {
                "summary": "获取爱安财总分信息",
                "description": "获取爱安财系统的总分信息",
                "operationId": "fetch_score_info_api_v1_aac_fetch_score_info_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthmeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ScoreInfoResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/provider__loveac__authme__AuthmeResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ],
                                    "title": "Response Fetch Score Info Api V1 Aac Fetch Score Info Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/aac/fetch_score_list": {
            "post": {
                "summary": "获取爱安财分数明细列表",
                "description": "获取爱安财系统的分数明细列表",
                "operationId": "fetch_score_list_api_v1_aac_fetch_score_list_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthmeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ScoreListResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/provider__loveac__authme__AuthmeResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ],
                                    "title": "Response Fetch Score List Api V1 Aac Fetch Score List Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuthmeRequest": {
                "properties": {
                    "token": {
                        "type": "string",
                        "title": "Token"
                    }
                },
                "type": "object",
                "required": [
                    "token"
                ],
                "title": "AuthmeRequest"
            },
...
            "ErrorResponse": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "description": "状态码",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "提示信息",
                        "default": "成功"
                    },
                    "data": {
                        "type": "null",
                        "title": "Data",
                        "description": "响应数据"
                    }
                },
                "type": "object",
                "title": "ErrorResponse",
                "description": "专用错误响应模型"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LoveACScoreCategory": {
                "properties": {
                    "ID": {
                        "type": "string",
                        "title": "Id",
                        "default": ""
                    },
                    "ShowNum": {
                        "type": "integer",
                        "title": "Shownum",
                        "default": 0
                    },
                    "TypeName": {
                        "type": "string",
                        "title": "Typename",
                        "default": ""
                    },
                    "TotalScore": {
                        "type": "number",
                        "title": "Totalscore",
                        "default": 0.0
                    },
                    "children": {
                        "items": {
                            "$ref": "#/components/schemas/LoveACScoreItem"
                        },
                        "type": "array",
                        "title": "Children",
                        "default": []
                    }
                },
                "type": "object",
                "title": "LoveACScoreCategory",
                "description": "爱安财分数类别"
            },
            "LoveACScoreInfo": {
                "properties": {
                    "TotalScore": {
                        "type": "number",
                        "title": "Totalscore",
                        "default": 0.0
                    },
                    "IsTypeAdopt": {
                        "type": "boolean",
                        "title": "Istypeadopt",
                        "default": false
                    },
                    "TypeAdoptResult": {
                        "type": "string",
                        "title": "Typeadoptresult",
                        "default": ""
                    }
                },
                "type": "object",
                "title": "LoveACScoreInfo",
                "description": "爱安财总分信息"
            },
            "LoveACScoreItem": {
                "properties": {
                    "ID": {
                        "type": "string",
                        "title": "Id",
                        "default": ""
                    },
                    "Title": {
                        "type": "string",
                        "title": "Title",
                        "default": ""
                    },
                    "TypeName": {
                        "type": "string",
                        "title": "Typename",
                        "default": ""
                    },
                    "UserNo": {
                        "type": "string",
                        "title": "Userno",
                        "default": ""
                    },
                    "Score": {
                        "type": "number",
                        "title": "Score",
                        "default": 0.0
                    },
                    "AddTime": {
                        "type": "string",
                        "title": "Addtime",
                        "default": ""
                    }
                },
                "type": "object",
                "title": "LoveACScoreItem",
                "description": "爱安财分数明细条目"
            },
...
            "ScoreInfoResponse": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "description": "状态码",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "提示信息",
                        "default": "成功"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/LoveACScoreInfo"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "响应数据"
                    }
                },
                "type": "object",
                "title": "ScoreInfoResponse",
                "description": "爱安财总分信息响应"
            },
            "ScoreListResponse": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "description": "状态码",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "提示信息",
                        "default": "成功"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/LoveACScoreCategory"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Data",
                        "description": "响应数据"
                    }
                },
                "type": "object",
                "title": "ScoreListResponse",
                "description": "爱安财分数明细列表响应"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "provider__loveac__authme__AuthmeResponse": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "title": "AuthmeResponse"
            },
...
        }
    }
}